requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
sqlite3  # já incluso no Python padrão
dateparser==1.2.0
requests==2.31.0